Refactors application into a multi-page structure
Separates chat functionality into its own page and introduces a dedicated memory page for displaying user memories.

Updates the main application logic to improve navigation and user experience by utilizing Streamlit's native multi-page features.

Enhances client initialization and configuration management for improved maintainability.

Relates to task Phase 8.
